requirements.txt
smtplib
imaplib
email
google-generativeai